Introduce test site dispatch in dao_setup
